Updates database entities to reflect new database schema

+++ PEClient/AspNetUsers.cs
@@ -33,7 +33,7 @@ namespace PEClient
         public int AccessFailedCount { get; set; }
         public string UserName { get; set; }
         public string FullName { get; set; }
-        public decimal UserId { get; set; }
+        public int UserId { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<AspNetUserRoles> AspNetUserRoles { get; set; }
