+++ PEClient/spAspNetUsers_GetAllStudents_Result.cs
@@ -15,6 +15,6 @@ namespace PEClient
     {
         public string FullName { get; set; }
         public string UserName { get; set; }
-        public decimal UserId { get; set; }
+        public int UserId { get; set; }
     }
 }
