+++ PEClient/tblTeamUsers.cs
@@ -14,9 +14,9 @@ namespace PEClient
     
     public partial class tblTeamUsers
     {
-        public decimal TeamUserId { get; set; }
-        public decimal TeamId { get; set; }
-        public decimal UserId { get; set; }
+        public int TeamUserId { get; set; }
+        public int TeamId { get; set; }
+        public int UserId { get; set; }
     
         public virtual tblTeams tblTeams { get; set; }
     }
